requirements.txt
Flask
Jinja2
pytest
pyyaml
google-cloud-aiplatform
